main.tsは事実上loginのため、loginに変更
