+++ apps/frontend/src/pages/login/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./authService";
